minor fixes

+++ create-express-app.js
@@ -8,6 +8,7 @@ const boxen = require('boxen');
 const ora = require('ora');
 const execa = require('execa');
 const logto = require('logto');
+const chalk = require('chalk');
 const { modifyPackageJson, removeDir } = require('./scripts/filehandling');
 const { performChecks } = require('./scripts/checks');
 const { getTemplateList } = require('./scripts/templates');
@@ -16,23 +17,31 @@ let metadata = {
     pkgmanager: "",
     template: "",
     projectname: "",
-    pkgmoption: ""  //package manager option(--cwd/--prefix)
+    pkgmoption: "",  //package manager option(--cwd/--prefix)
+    debug: ""
 }
+// display banner
+console.log(boxen('create-express-app v0.0.1', { padding: 0, margin: 0, borderColor: "magentaBright", borderStyle: "round" }));
+
+// check if debug mode is enabled
+const arg = process.argv.splice(2)[0];
+if (arg === "--debug" || arg === '-d') {
+    metadata.debug = true;
+    console.log(emojis.get('bug'), chalk.cyan('debug mode enabled'));
+};
 // starting new logfile instance
 const logdir = path.join(os.homedir(), ".create-express-app-data", "logs");
 const logfile = `${getDateTime()}.log`;
 const logger = new logto({ dir: logdir, file: logfile });
-// display banner
-console.log(boxen('create-express-app v0.0.1', { padding: 0, margin: 0, borderColor: "magentaBright", borderStyle: "round" }));
 
 (async () => {
     // perform checks before proceeding
     let spinner = ora('performing checks').start();
     let result = await performChecks().catch(err => {
         // log error to log file
-        logger.log(err);
-        spinner.fail(err.message);
-        console.log(`log file can be found in ${logger.logfile}`);
+        logger.log(err.stack);
+        spinner.fail(`${chalk.redBright(err.message)}`);
+        console.log(`log file can be found at ${chalk.cyan(logger.logfile)}`);
         process.exit(1);
     });
     Object.assign(metadata, result);
@@ -42,9 +51,9 @@ console.log(boxen('create-express-app v0.0.1', { padding: 0, margin: 0, borderCo
     spinner.start();
     const templateList = await getTemplateList().catch(err => {
         // log error to log file
-        logger.log(err);
-        spinner.fail("some error occured while fetching templates. Make sure you have active internet connection");
-        console.log(`log file can be found in ${logger.logfile}`);
+        logger.log(err.stack);
+        spinner.fail(`${chalk.redBright("some error occured while fetching templates. Make sure you have active internet connection\n")}`);
+        console.log(`log file can be found at ${chalk.cyan(logger.logfile)}`);
         process.exit(1);
     });
     spinner.succeed("fetched templates\n");
@@ -76,9 +85,9 @@ console.log(boxen('create-express-app v0.0.1', { padding: 0, margin: 0, borderCo
         console.log(`${emojis.get("open_file_folder")} ${metadata.projectname}  created!`);
     }).catch(err => {
         // log error to log file
-        logger.log(err);
-        console.log(`${emojis.get('x')}${err.message}`);
-        console.log(`log file can be found in ${logger.logfile}`);
+        logger.log(err.stack);
+        console.log(`${emojis.get('x')}${chalk.redBright(err.message)}`);
+        console.log(`log file can be found at ${chalk.cyan(logger.logfile)}`);
         process.exit(1);
     });
     // copy selected template from app data to project dir
@@ -90,12 +99,12 @@ console.log(boxen('create-express-app v0.0.1', { padding: 0, margin: 0, borderCo
         })
         .catch(err => {
             // log error to log file
-            logger.log(err);
+            logger.log(err.stack);
             // remove project directory
             removeDir(metadata.projectname);
             //then log error
-            console.log(`${emojis.get('x')} ${err.message}`);
-            console.log(`log file can be found in ${logger.logfile}`);
+            console.log(`${emojis.get('x')}${chalk.redBright(err.message)}`);
+            console.log(`log file can be found at ${chalk.cyan(logger.logfile)}`);
             process.exit(1);
         })
     //make changes to package.json in project directory
@@ -103,12 +112,12 @@ console.log(boxen('create-express-app v0.0.1', { padding: 0, margin: 0, borderCo
         console.log(`${emojis.get('memo')} modified package.json\n`);
     }).catch(err => {
         // log error to log file
-        logger.log(err);
+        logger.log(err.stack);
         // remove project directory
         removeDir(metadata.projectname);
         //then log error
-        console.log(`${emojis.get('x')} ${err.message}`);
-        console.log(`log file can be found in ${logger.logfile}`);
+        console.log(`${emojis.get('x')}${chalk.redBright(err.message)}`);
+        console.log(`log file can be found at ${chalk.cyan(logger.logfile)}`);
         process.exit(1);
     });
     // perform installs as per the package manager selected in checks
@@ -116,12 +125,12 @@ console.log(boxen('create-express-app v0.0.1', { padding: 0, margin: 0, borderCo
     spinner.start();
     await execa(metadata.pkgmanager, ['install', metadata.pkgmoption, `${metadata.projectname}/`]).catch(err => {
         // log error to log file
-        logger.log(err);
+        logger.log(err.stack);
         // remove project directory
         removeDir(metadata.projectname);
         //then log error
-        sp.fail(err.message);
-        console.log(`log file can be found in ${logger.logfile}`);
+        sp.fail(chalk.redBright(err.message));
+        console.log(`log file can be found in ${chalk.cyan(logger.logfile)}`);
         process.exit(1);
     });
     spinner.succeed(`${emojis.get('package')} node modules installed!`);
@@ -130,12 +139,12 @@ console.log(boxen('create-express-app v0.0.1', { padding: 0, margin: 0, borderCo
     spinner.start();
     await execa('git', ['init', `${metadata.projectname}/`]).catch(err => {
         // log error to log file
-        logger.log(err);
+        logger.log(err.stack);
         // remove project directory
         removeDir(metadata.projectname);
         //then log error
-        spinner.fail(err.message);
-        console.log(`log file can be found in ${logger.logfile}`);
+        sp.fail(chalk.redBright(err.message));
+        console.log(`log file can be found in ${chalk.cyan(logger.logfile)}`);
         process.exit(1);
     });
     spinner.succeed(`${emojis.get('octopus')} git repo initialized`);
@@ -145,5 +154,10 @@ console.log(boxen('create-express-app v0.0.1', { padding: 0, margin: 0, borderCo
     let rp = getRandomPhrase();
     console.log(boxen(rp, { padding: 0, margin: 0, borderColor: "yellow", borderStyle: "round" }));
     // reached here => no errors => no need of log files. remove em
-    removeDir(logdir);
+    // if debug mode then don't delete log
+    if (!metadata.debug) {
+        removeDir(logdir);
+        return;
+    };
+    console.log(`log file can be found at`, chalk.cyan(logger.logfile));
 })();
